--- src/components/models/hero-models/HeroExperience.tsx
@@ -1,30 +1,0 @@
-import { Canvas } from "@react-three/fiber";
-import { OrbitControls } from "@react-three/drei";
-import { useMediaQuery } from "react-responsive";
-import HeroLights from "./HeroLights";
-import Particles from "./Particles";
-import { Room } from "./Room";
-
-const HeroExperience = () => {
-  const isTablet: boolean = useMediaQuery({ query: "(max-width: 1224px)" });
-  const isMobile: boolean = useMediaQuery({ query: "(max-width: 768px)" });
-  return (
-    <Canvas camera={{ position: [0, 0, 0], fov: 45 }}>
-      <OrbitControls
-        enablePan={false}
-        enableZoom={!isTablet}
-        maxDistance={20}
-        minDistance={5}
-        maxPolarAngle={Math.PI / 2}
-        minPolarAngle={Math.PI / 5}
-      />
-      <HeroLights />
-      <Particles count={100} />
-      <group scale={isMobile ? 0.7 : 1} position={[0, -3.5, 0]} rotation={[0, -Math.PI / 4, 0]}>
-        <Room />
-      </group>
-    </Canvas>
-  );
-};
-
-export default HeroExperience;
