--- src/components/models/hero-models/Particles.tsx
@@ -1,68 +1,0 @@
-import { useRef, useMemo } from "react";
-import { useFrame } from "@react-three/fiber";
-import * as THREE from "three";
-
-interface ParticlesProps {
-  count?: number;
-}
-
-interface Particle {
-  position: [number, number, number];
-  speed: number;
-}
-
-const Particles = ({ count = 200 }: ParticlesProps) => {
-  const mesh = useRef<THREE.Points | null>(null);
-
-  const particles = useMemo<Particle[]>(() => {
-    const temp: Particle[] = [];
-    for (let i = 0; i < count; i++) {
-      temp.push({
-        position: [
-          (Math.random() - 0.5) * 10,
-          Math.random() * 10 + 5, // higher starting point
-          (Math.random() - 0.5) * 10,
-        ],
-        speed: 0.005 + Math.random() * 0.001,
-      });
-    }
-    return temp;
-  }, [count]);
-
-  useFrame(() => {
-    if (mesh.current) {
-      const positions = mesh.current.geometry.attributes.position;
-      if (positions && positions.array) {
-        const posArray = positions.array as Float32Array;
-        for (let i = 0; i < count; i++) {
-          let y = posArray[i * 3 + 1];
-          y -= particles[i].speed;
-          if (y < -2) y = Math.random() * 10 + 5;
-          posArray[i * 3 + 1] = y;
-        }
-        positions.needsUpdate = true;
-      }
-    }
-  });
-
-  const positions = useMemo(() => {
-    const pos = new Float32Array(count * 3);
-    particles.forEach((p, i) => {
-      pos[i * 3] = p.position[0];
-      pos[i * 3 + 1] = p.position[1];
-      pos[i * 3 + 2] = p.position[2];
-    });
-    return pos;
-  }, [particles, count]);
-
-  return (
-    <points ref={mesh}>
-      <bufferGeometry>
-        <bufferAttribute attach="attributes-position" args={[positions, 3]} />
-      </bufferGeometry>
-      <pointsMaterial color="#ffffff" size={0.05} transparent opacity={0.9} depthWrite={false} />
-    </points>
-  );
-};
-
-export default Particles;
