+++ src/sections/Footer.tsx
@@ -2,7 +2,7 @@ import { socialImgs } from "@/constants";
 
 const Footer = () => {
   return (
-    <footer className="footer">
+    <footer className="relative footer z-10">
       <div className="footer-container">
         <div className="flex flex-col justify-center">
           <p>Terms & Conditions</p>
